Remove unused files and change file structure
