+++ pavlovStanislav/src/components/customs/text/MediumText.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import {Text} from 'react-native'
-import {CustomText} from '../../../constants/types'
+import {ICustomText} from '../../../interfaces/custom-text.interface'
 
-const MediumText: React.FC<CustomText> = ({style, children}) => {
+const MediumText: React.FC<ICustomText> = ({style, children}) => {
   return (
     <Text style={[style, {fontFamily: 'Montserrat-Medium'}]}>{children}</Text>
   )
