+++ pavlovStanislav/src/components/customs/text/SemiBoldText.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import {Text} from 'react-native'
-import {CustomText} from '../../../constants/types'
+import {ICustomText} from '../../../interfaces/custom-text.interface'
 
-const SemiBoldText: React.FC<CustomText> = ({style, children}) => {
+const SemiBoldText: React.FC<ICustomText> = ({style, children}) => {
   return (
     <Text style={[style, {fontFamily: 'Montserrat-SemiBold'}]}>{children}</Text>
   )
