--- pavlovStanislav/src/constants/colors.ts
@@ -1,10 +1,0 @@
-export const colors = {
-    white: '#FFF',
-    black: '#111',
-    gray: '#F2F0F0',
-    darkGray: '#787878',
-    purple: '#424874',
-    darkPurple: '#3c415e',
-    lightLilac: '#DCD6F7',
-    orange: '#FF7A00',
-}
