+++ pavlovStanislav/src/constants/dimensions.const.ts
@@ -1,0 +1,6 @@
+import {Dimensions} from 'react-native'
+
+export const dimensions = {
+  width: Dimensions.get('window').width,
+  height: Dimensions.get('window').height,
+}
