--- pavlovStanislav/src/constants/dimensions.ts
@@ -1,6 +1,0 @@
-import { Dimensions } from 'react-native'
-
-export const dimensions = {
-    width: Dimensions.get('window').width,
-    height: Dimensions.get('window').height,
-}
