+++ pavlovStanislav/src/interfaces/custom-text.interface.ts
@@ -1,0 +1,7 @@
+import {ReactNode} from 'react'
+import {StyleProp, TextStyle} from 'react-native'
+
+export interface ICustomText {
+  style?: StyleProp<TextStyle>
+  children: ReactNode
+}
